move CorsSupport and ERR into utils package
