+++ jvm/src/main/scala/easydocs/routes/requests/EndpointCreateRequest.scala
@@ -5,9 +5,9 @@ import java.util.UUID
 import com.sksamuel.elastic4s.ElasticClient
 import com.sksamuel.elastic4s.ElasticDsl._
 import com.sksamuel.elastic4s.source.ObjectSource
-import easydocs.ERR
 import easydocs.models.{ESSite, ESEndpoint}
 import easydocs.routes.responses.{EndpointListResponse, BooleanResponse}
+import easydocs.utils.ERR
 import spray.json.DefaultJsonProtocol._
 
 import scala.concurrent.{Future, ExecutionContext}
